+++ eslint.config.mjs
@@ -1,16 +1,19 @@
 // @ts-check
-import withNuxt from './.nuxt/eslint.config.mjs'
+import withNuxt from './.nuxt/eslint.config.mjs';
 
 export default withNuxt(
-  // ...Custom flat configs append after nuxt's configs
- 
-) .prepend(
-  // ...Prepend some flat configs in front
-)
-.override('nuxt/typescript', {
-  rules: {
-    // ...Override rules, for example:
-    // '@typescript-eslint/ban-types': 'off'
-    
-  }
-})
+	// ...Custom flat configs append after nuxt's configs
+	{
+		files: ['**/*.ts', '**/*.tsx'],
+		rules: {
+			'no-trailing-spaces': ['warn'],
+			'prefer-promise-reject-errors': 'off',
+			'vue/no-v-html': 'off',
+			'vue/multi-word-component-names': ['error', {
+				ignores: ['error', 'Error'],
+			}],
+
+		},
+
+	},
+);