+++ nuxt.config.ts
@@ -1,15 +1,34 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
-  devtools: { enabled: true },
-  modules: ['@nuxt/devtools', '@nuxt/eslint'],
-  eslint: {
-    // options here
-    config: {
-      stylistic: true,
-    },
-  },
-  typescript: {
-    typeCheck: true,
-  },
+	devtools: { enabled: true },
+	app: {
+		pageTransition: { name: 'page', mode: 'out-in' },
+	},
+	css: ['~/assets/styles/main.scss'],
+	modules: [
+		'@nuxt/devtools',
+		'@nuxt/eslint',
+		'@nuxtjs/tailwindcss',
+		'@nuxtjs/color-mode',
+		'@pinia/nuxt',
+		'@nuxt/fonts',
+	],
+	eslint: {
+		// options here
+		config: {
+			stylistic: {
+				indent: 'tab',
+				quotes: 'single',
+				semi: true,
+			},
+		},
 
-})
+	},
+	pinia: {
+		storesDirs: ['./stores/**', './custom-folder/stores/**'],
+	},
+	typescript: {
+		typeCheck: true,
+	},
+
+});