+++ pages/index.vue
@@ -1,0 +1,31 @@
+<script lang="ts" setup>
+
+</script>
+
+<template>
+	<div>
+		<h1>Home Page</h1>
+		<NuxtLink
+			to="/about"
+			class="custom-link"
+			active-class="active-link"
+		>
+			About page
+		</NuxtLink>
+	</div>
+</template>
+
+<style scoped>
+	h1 {
+		font-family: "Inter", sans-serif;
+	}
+  .custom-link {
+    text-decoration: underline;
+    /* Additional styling if needed */
+  }
+  .active-link {
+    /* Your active link styling here */
+    font-weight: bold;
+    color: #007bff;
+  }
+</style>