+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## 1.3.0 - 2023-Jan-10
 
 ### Added
-- Added completed feature
+- Added App.vue with Layout and Page transition
 
 ### Changed
 2024-May-15
+- Added App.vue
+- Added Nuxt Layout and Nuxt Page with transition
+
+2024-May-13
 - Jumping to Version 3, pushing to master
 - troubleshooting ESlint
 
@@ -42,3 +46,5 @@
 
 ### Security
 - Security-related changes.
+
+